refactor(web): share module rate Sparkline component
